created test_array.py for all unittests in array/ module
