+++ array/move_zeros_to_end.py
@@ -6,7 +6,6 @@ preserving the order of the other elements.
 
 The time complexity of the below algorithm is O(n).
 """
-import unittest
 
 
 def move_zeros(array):
@@ -21,19 +20,3 @@ def move_zeros(array):
     
     result.extend([0] * zeros)
     return result
-
-
-class TestSuite(unittest.TestCase):
-
-    def test_move_zeros(self):
-
-        self.assertListEqual(move_zeros([False, 1, 0, 1, 2, 0, 1, 3, "a"]),
-                             [False, 1, 1, 2, 1, 3, "a", 0, 0])
-
-        self.assertListEqual(move_zeros([0, 34, 'rahul', [], None, 0, True, 0]),
-                             [34, 'rahul', [], None, True, 0, 0, 0])
-
-
-if __name__ == '__main__':
-
-    unittest.main()
