+++ array/plus_one.py
@@ -5,7 +5,6 @@ plus one to the number.
 The digits are stored such that the most significant
 digit is at the head of the list.
 """
-import unittest
 
 
 def plus_one(digits):
@@ -47,35 +46,3 @@ def plus_one_v3(num_arr):
         if num_arr[idx]:
             return num_arr
     return [1] + num_arr
-
-
-class TestSuite(unittest.TestCase):
-
-    def test_plus_one(self):
-
-        self.assertListEqual(plus_one([0]), [1])
-        self.assertListEqual(plus_one([9]), [1, 0])
-        self.assertListEqual(plus_one([1, 0, 9]), [1, 1, 0])
-        self.assertListEqual(plus_one([9, 9, 8, 0, 0, 9]), [9, 9, 8, 0, 1, 0])
-        self.assertListEqual(plus_one([9, 9, 9, 9]), [1, 0, 0, 0, 0])
-
-    def test_plus_one_v2(self):
-
-        self.assertListEqual(plus_one_v2([0]), [1])
-        self.assertListEqual(plus_one_v2([9]), [1, 0])
-        self.assertListEqual(plus_one_v2([1, 0, 9]), [1, 1, 0])
-        self.assertListEqual(plus_one_v2([9, 9, 8, 0, 0, 9]), [9, 9, 8, 0, 1, 0])
-        self.assertListEqual(plus_one_v2([9, 9, 9, 9]), [1, 0, 0, 0, 0])
-
-    def test_plus_one_v3(self):
-
-        self.assertListEqual(plus_one_v3([0]), [1])
-        self.assertListEqual(plus_one_v3([9]), [1, 0])
-        self.assertListEqual(plus_one_v3([1, 0, 9]), [1, 1, 0])
-        self.assertListEqual(plus_one_v3([9, 9, 8, 0, 0, 9]), [9, 9, 8, 0, 1, 0])
-        self.assertListEqual(plus_one_v3([9, 9, 9, 9]), [1, 0, 0, 0, 0])
-
-
-if __name__ == '__main__':
-
-    unittest.main()
