+++ array/summary_ranges.py
@@ -4,7 +4,6 @@ return the summary of its ranges.
 
 For example, given [0, 1, 2, 4, 5, 7], return [(0, 2), (4, 5), (7, 7)].
 """
-import unittest
 
 
 def summary_ranges(array):
@@ -26,20 +25,3 @@ def summary_ranges(array):
             res.append((num, num))
         i += 1
     return res
-
-
-class TestSuite(unittest.TestCase):
-
-    def test_summary_ranges(self):
-
-        self.assertListEqual(summary_ranges([0, 1, 2, 4, 5, 7]),
-                             [(0, 2), (4, 5), (7, 7)])
-        self.assertListEqual(summary_ranges([-5, -4, -3, 1, 2, 4, 5, 6]),
-                             [(-5, -3), (1, 2), (4, 6)])
-        self.assertListEqual(summary_ranges([-2, -1, 0, 1, 2]),
-                             [(-2, 2)])
-
-
-if __name__ == '__main__':
-
-    unittest.main()
