+++ array/test_array.py
@@ -2,6 +2,15 @@ from delete_nth import delete_nth, delete_nth_naive
 from flatten import flatten, flatten_iter
 from garage import garage
 from josephus_problem import josephus
+from longest_non_repeat import longest_non_repeat, longest_non_repeat_two
+from merge_intervals import Interval, merge_v2
+from missing_ranges import missing_ranges
+from move_zeros_to_end import move_zeros
+from plus_one import plus_one, plus_one_v2, plus_one_v3
+from rotate_array import rotate_v1, rotate_v2, rotate_v3
+from summary_ranges import summary_ranges
+from three_sum import three_sum
+from two_sum import two_sum
 
 import unittest
 
@@ -99,5 +108,171 @@ class TestFlatten(unittest.TestCase):
         self.assertRaises(StopIteration, next, flattened)
 
 
+class TestGarage(unittest.TestCase):
+
+    def test_garage(self):
+
+        initial = [1, 2, 3, 0, 4]
+        final = [0, 3, 2, 1, 4]
+        steps, seq = garage(initial, final)
+
+        self.assertEqual(steps, 4)
+        self.assertListEqual(seq, [[0, 2, 3, 1, 4], [2, 0, 3, 1, 4], [2, 3, 0, 1, 4], [0, 3, 2, 1, 4]])
+
+
+class TestLongestNonRepeat(unittest.TestCase):
+
+    def test_longest_non_repeat(self):
+
+        string = "abcabcbb"
+        self.assertEqual(longest_non_repeat(string), 3)
+
+        string = "bbbbb"
+        self.assertEqual(longest_non_repeat(string), 1)
+
+        string = "pwwkew"
+        self.assertEqual(longest_non_repeat(string), 3)
+
+    def test_longest_non_repeat_two(self):
+
+        string = "abcabcbb"
+        self.assertEqual(longest_non_repeat_two(string), 3)
+
+        string = "bbbbb"
+        self.assertEqual(longest_non_repeat_two(string), 1)
+
+        string = "pwwkew"
+        self.assertEqual(longest_non_repeat_two(string), 3)
+
+
+class TestMergeInterval(unittest.TestCase):
+
+    def test_merge(self):
+        interval_list = [[1, 3], [2, 6], [8, 10], [15, 18]]
+        intervals = [Interval(i[0], i[1]) for i in interval_list]
+        merged_intervals = Interval.merge(intervals)
+        self.assertEqual(
+            merged_intervals,
+            [Interval(1, 6), Interval(8, 10), Interval(15, 18)]
+        )
+
+    def test_merge_v2(self):
+        interval_list = [[1, 3], [2, 6], [8, 10], [15, 18]]
+        merged_intervals = merge_v2(interval_list)
+        self.assertEqual(
+            merged_intervals,
+            [[1, 6], [8, 10], [15, 18]]
+        )
+
+
+class TestMissingRanges(unittest.TestCase):
+
+    def test_missing_ranges(self):
+
+        arr = [3, 5, 10, 11, 12, 15, 19]
+
+        self.assertListEqual(missing_ranges(arr, 0, 20),
+                             [(0, 2), (4, 4), (6, 9), (13, 14), (16, 18), (20, 20)])
+
+        self.assertListEqual(missing_ranges(arr, 6, 100),
+                             [(6, 9), (13, 14), (16, 18), (20, 100)])
+
+
+class TestMoveZeros(unittest.TestCase):
+
+    def test_move_zeros(self):
+
+        self.assertListEqual(move_zeros([False, 1, 0, 1, 2, 0, 1, 3, "a"]),
+                             [False, 1, 1, 2, 1, 3, "a", 0, 0])
+
+        self.assertListEqual(move_zeros([0, 34, 'rahul', [], None, 0, True, 0]),
+                             [34, 'rahul', [], None, True, 0, 0, 0])
+
+
+class TestPlusOne(unittest.TestCase):
+
+    def test_plus_one(self):
+
+        self.assertListEqual(plus_one([0]), [1])
+        self.assertListEqual(plus_one([9]), [1, 0])
+        self.assertListEqual(plus_one([1, 0, 9]), [1, 1, 0])
+        self.assertListEqual(plus_one([9, 9, 8, 0, 0, 9]), [9, 9, 8, 0, 1, 0])
+        self.assertListEqual(plus_one([9, 9, 9, 9]), [1, 0, 0, 0, 0])
+
+    def test_plus_one_v2(self):
+
+        self.assertListEqual(plus_one_v2([0]), [1])
+        self.assertListEqual(plus_one_v2([9]), [1, 0])
+        self.assertListEqual(plus_one_v2([1, 0, 9]), [1, 1, 0])
+        self.assertListEqual(plus_one_v2([9, 9, 8, 0, 0, 9]), [9, 9, 8, 0, 1, 0])
+        self.assertListEqual(plus_one_v2([9, 9, 9, 9]), [1, 0, 0, 0, 0])
+
+    def test_plus_one_v3(self):
+
+        self.assertListEqual(plus_one_v3([0]), [1])
+        self.assertListEqual(plus_one_v3([9]), [1, 0])
+        self.assertListEqual(plus_one_v3([1, 0, 9]), [1, 1, 0])
+        self.assertListEqual(plus_one_v3([9, 9, 8, 0, 0, 9]), [9, 9, 8, 0, 1, 0])
+        self.assertListEqual(plus_one_v3([9, 9, 9, 9]), [1, 0, 0, 0, 0])
+
+
+class TestRotateArray(unittest.TestCase):
+
+    def test_rotate_v1(self):
+
+        self.assertListEqual(rotate_v1([1, 2, 3, 4, 5, 6, 7], k=3), [5, 6, 7, 1, 2, 3, 4])
+        self.assertListEqual(rotate_v1([1, 2, 3, 4, 5, 6, 7], k=1), [7, 1, 2, 3, 4, 5, 6])
+        self.assertListEqual(rotate_v1([1, 2, 3, 4, 5, 6, 7], k=7), [1, 2, 3, 4, 5, 6, 7])
+        self.assertListEqual(rotate_v1([1, 2], k=111), [2, 1])
+
+    def test_rotate_v2(self):
+
+        self.assertListEqual(rotate_v2([1, 2, 3, 4, 5, 6, 7], k=3), [5, 6, 7, 1, 2, 3, 4])
+        self.assertListEqual(rotate_v2([1, 2, 3, 4, 5, 6, 7], k=1), [7, 1, 2, 3, 4, 5, 6])
+        self.assertListEqual(rotate_v2([1, 2, 3, 4, 5, 6, 7], k=7), [1, 2, 3, 4, 5, 6, 7])
+        self.assertListEqual(rotate_v2([1, 2], k=111), [2, 1])
+
+    def test_rotate_v3(self):
+
+        self.assertListEqual(rotate_v3([1, 2, 3, 4, 5, 6, 7], k=3), [5, 6, 7, 1, 2, 3, 4])
+        self.assertListEqual(rotate_v3([1, 2, 3, 4, 5, 6, 7], k=1), [7, 1, 2, 3, 4, 5, 6])
+        self.assertListEqual(rotate_v3([1, 2, 3, 4, 5, 6, 7], k=7), [1, 2, 3, 4, 5, 6, 7])
+        self.assertListEqual(rotate_v3([1, 2], k=111), [2, 1])
+
+
+class TestSummaryRanges(unittest.TestCase):
+
+    def test_summary_ranges(self):
+
+        self.assertListEqual(summary_ranges([0, 1, 2, 4, 5, 7]),
+                             [(0, 2), (4, 5), (7, 7)])
+        self.assertListEqual(summary_ranges([-5, -4, -3, 1, 2, 4, 5, 6]),
+                             [(-5, -3), (1, 2), (4, 6)])
+        self.assertListEqual(summary_ranges([-2, -1, 0, 1, 2]),
+                             [(-2, 2)])
+
+
+class TestThreeSum(unittest.TestCase):
+
+    def test_three_sum(self):
+
+        self.assertSetEqual(three_sum([-1, 0, 1, 2, -1, -4]),
+                            {(-1, 0, 1), (-1, -1, 2)})
+
+        self.assertSetEqual(three_sum([-1, 3, 1, 2, -1, -4, -2]),
+                            {(-4, 1, 3), (-2, -1, 3), (-1, -1, 2)})
+
+
+class TestSuite(unittest.TestCase):
+
+    def test_two_sum(self):
+
+        self.assertTupleEqual((0, 2), two_sum([2, 11, 7, 9], target=9))
+        self.assertTupleEqual((0, 3), two_sum([-3, 5, 2, 3, 8, -9], target=0))
+
+        self.assertIsNone(two_sum([-3, 5, 2, 3, 8, -9], target=6))
+
+
 if __name__ == '__main__':
+    
     unittest.main()
